requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
nltk==3.8.1
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
pypdf2==3.0.1
python-docx==1.1.0
textract==1.6.5
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.29.0 